Add files via upload
Practical done in university C++ module. We were expected to output a menu, allow the user to choose an item, and use a switch statement accordingly to navigate to different calculations. Option 1:  calculate and display the circumference and area of a circle, based on the radius provided by a user. Option 2: convert temperature values between Celsius and fahrenheit. Option 3: check whether a given year is a leap year.