Add files via upload
A C++ program that simulates recording and analysing daily temperatures over a week.
The program generates random temperature values for each day of the week in the
range of 15.0 to 35.0 degrees Celsius. After generating the temperatures, the program
displays the temperatures and identifies the highest and lowest temperatures recorded during
the week.