Add files via upload
This was my introduction to structures. This is a simple C++ program that stores and displays personal information using a structured. The program should capture a person’s South African ID number (13-digit), calculate their age from the ID number, determine if the person is male or female, and then display this information