Add files via upload
This program is an "Adopt a Pet" program. Based on the visitor’s decision,
the program should display the pet type, and the randomly generated adoption fee, and ask if they would like to adopt another pet. The program should continue until the user chooses not to adopt another pet.